Minor fixes and cleaning code, playing videos again.
