+++ config.json
@@ -1,10 +1,12 @@
-{	"fps": "25", 
+{
+    "fps": "25", 
     "resolution": "hd", 
     "playout_mode": "direct", 
     "playlist_maxlength": "4 hours", 
     "scheduled_playlist_maxlength": "04:00:00", 
     "timer_interval": "1000",
     "black": "../../images/black.png",
+    "reload_timer_diff": "20000",
     "playlist_server": "mongo",
     "mvcp_server": "melted"
 }
