+++ utils.js
@@ -1,3 +1,5 @@
+var moment = require('moment');
+
 exports = module.exports = {
     getXmlFileNameFromClip: function(clip) {
         return clip.playlist_id + "-" + clip.id + ".xml";
@@ -37,5 +39,33 @@ exports = module.exports = {
     
     getCurrentPosFromClip: function(actualFrame, totalFrames) {
         return parseFloat(actualFrame / totalFrames);
+    },
+
+    convertFramesToSeconds: function ( frames, fps ) {
+        return frames/fps;
+    },
+
+    convertLengthToMilliseconds: function ( frames ) {
+        var m = moment( frames, "HH:mm:ss.SS");
+        return m.hours()*60*60*1000 + m.minutes()*60*1000 + m.seconds()*1000 + m.milliseconds();
+    },
+    
+    convertFramesToMilliseconds: function ( frames, fps ) {
+		if (fps+""=="NaN" || fps==undefined || fps===false || fps==0) {
+            var m = moment( frames, "HH:mm:ss.SS");
+            return m.hours()*60*60*1000 + m.minutes()*60*1000 + m.seconds()*1000 + m.milliseconds();
+ 		}
+        return frames * 1000.0 / (1.0 * fps);
+    },
+    
+    convertUnixToDate:  function ( unix_timestamp ) {
+        //var date = new Date(unix_timestamp*1000);     
+        var date = new moment(unix_timestamp);
+        return date.format("hh:mm:ss");
+    },
+
+    convertDateToUnix:  function ( date_timestamp ) {        
+        var date = new moment(date_timestamp);
+        return date.unix();        
     }
 };
